实现文件块分配

+++ DB/Convertor.h
@@ -11,17 +11,17 @@ private:
 	virtual ~Conv64() {}
 
 public:
-	static int64_t sti(const string& strValue);
-	static string its(int64_t nValue);
+	static int sti(const string& strValue);
+	static string its(int nValue);
 	static string to_64(const string& src10, const int nMinLen = 0);
 	//10进制转化成的x进制长度，不足nMinLen长度的话高位补0
-	static string to_64(const int64_t src10, int nMinLen = 0);
+	static string to_64(const int src10, int nMinLen = 0);
 	//x进制转为10进制
 	static string to_10(const string& srcx, int nMinLen = 0);
 protected:
 	static string convMinLen(const string& src, int nMinLen);
 	//初始化，设置字符集
-	void init(const char* charSet);
+	void init(const char * charSet);
 private:
 	static char m_charSet[128];
 	static int m_nDigital;
